+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2024-06-29 (tagged 8.1.3.40)
 
  * Update to Onload 8.1.3.40
+ * Add Ubuntu `noble` flavor
 
 ## 2024-01-17 (tagged 8.1.2.26)